refactor(CHAOS-1204): break cyclic import — move STANDARD_FEATURES to licensing/registry
CodeQL flagged py/cyclic-import on the `models.licensing ↔ licensing.types`
pair (even through a lazy inline import, the structural cycle remained).

- New `licensing/registry.py` owns STANDARD_FEATURES and get_features_for_tier
- `licensing/types.py` gains FeatureCategory and TIER_ORDER (previously in models)
- `models/licensing.py` imports the enums + STANDARD_FEATURES from the canonical
  homes; keeps them as module-level back-compat re-exports for existing callers
- `alembic/versions/0006_seed_feature_flags.py` declares __all__ to silence
  CodeQL py/unused-global-variable on Alembic's introspected module globals

101 licensing/billing tests pass; pre-existing test_checkout failures remain.
